refactor: extract types (no changes)
