+++ src/html-to-epub.ts
@@ -27,6 +27,7 @@ import { visit } from "unist-util-visit";
 import uslug from "uslug";
 
 import { allowedAttributes, allowedXhtml11Tags } from "./constants-gen";
+import { EpubContent, EpubImage, EpubOptions } from "./html-to-epub.types";
 
 // UUID generation
 function uuid() {
@@ -38,60 +39,6 @@ function uuid() {
 
 // Current directory
 const __dirname = fileURLToPath(new URL(".", import.meta.url));
-
-export interface EpubContentOptions {
-  title: string;
-  data: string;
-  url?: string;
-  author?: Array<string> | string;
-  filename?: string;
-  excludeFromToc?: boolean;
-  beforeToc?: boolean;
-}
-
-export interface EpubOptions {
-  title: string;
-  description: string;
-  cover?: string;
-  publisher?: string;
-  author?: Array<string> | string;
-  tocTitle?: string;
-  appendChapterTitles?: boolean;
-  date?: string;
-  lang?: string;
-  css?: string;
-  fonts?: Array<string>;
-  content: Array<EpubContentOptions>;
-  customOpfTemplatePath?: string;
-  customNcxTocTemplatePath?: string;
-  customHtmlTocTemplatePath?: string;
-  customHtmlCoverTemplatePath?: string;
-  version?: number;
-  userAgent?: string;
-  verbose?: boolean;
-  tempDir?: string;
-}
-
-interface EpubContent {
-  id: string;
-  href: string;
-  title: string;
-  data: string;
-  url: string | null;
-  author: Array<string>;
-  filePath: string;
-  templatePath: string;
-  excludeFromToc: boolean;
-  beforeToc: boolean;
-}
-
-interface EpubImage {
-  id: string;
-  url: string;
-  dir: string;
-  mediaType: string;
-  extension: string;
-}
 
 export class EPub {
   uuid: string;
